Add files via upload
The project aims to predict customer churn or loss for a telecommunications company, which refers to clients leaving a company, and its significant role in other business domains such as insurance companies, banks, and ISPs. The two main reasons for customer churn are identified as the service provider no longer meeting the customer's needs and customers being offered better incentives by competing service providers. To build a churn prediction model, I used churn-80 and churn-20 datasets for training and testing. The project team also cleaned and analyzed the data exploratory to gain a better understanding of it. Preprocessed and transformed data were used to create model-readable formats. A decision tree classifier model was used for prediction and evaluation, and a parameter grid was utilized to determine the best decision tree classifier for the dataset. Through the project's development, various insights were gained into how churn is predicted using Spark. The project's results and model can assist businesses in determining which customers are likely to churn and making appropriate business decisions based on those results.